backend/pyproject.toml
[project]
name = "sql-agent-backend"
description = "FastAPI backend — natural language to SQL via LlamaIndex + Claude"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "anthropic>=0.28.0",
    "llama-index-core==0.14.12",
    "llama-index-llms-anthropic>=0.11.1",
    "llama-index-embeddings-fastembed>=0.2.0",
    "fastembed>=0.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
